pyproject.toml
[tool.poetry]
name = "nonebot_plugin_oddtext"
version = "0.2.0"
description = "Nonebot2 插件，用于抽象话等文本生成"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-oddtext"
repository = "https://github.com/noneplugin/nonebot-plugin-oddtext"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = { version = "^2.0.0", extras = ["fastapi"] }
rcnb = "^1.0.3"

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
